requirements.txt
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-api-python-client==2.102.0
slack-sdk==3.26.2
apscheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
